fix: change name. Refs #179
